AP-2599: Refactor HMRC workers to reduce cognitive complexity
